refactor: TestFactory 글로벌로 변경
